+++ display_examples/lib_jpeg_hw_accel/lib_jpeg_hw_accel.c
@@ -181,10 +181,10 @@ void APP_DisplayDecodedImage(void)
 {
     /* Step 10: Configure dispaly layer configuration. */
     fbInfo.pixelFormat = APP_FB_FORMAT;
-    fbInfo.width       = cinfo.output_width;
-    fbInfo.height      = cinfo.output_height;
-    fbInfo.startX      = (DEMO_PANEL_WIDTH - cinfo.output_width) / 2U;
-    fbInfo.startY      = (DEMO_PANEL_HEIGHT - cinfo.output_height) / 2U;
+    fbInfo.width       = (uint16_t)cinfo.output_width;
+    fbInfo.height      = (uint16_t)cinfo.output_height;
+    fbInfo.startX      = (DEMO_PANEL_WIDTH - (uint16_t)cinfo.output_width) / 2U;
+    fbInfo.startY      = (DEMO_PANEL_HEIGHT - (uint16_t)cinfo.output_height) / 2U;
     fbInfo.strideBytes = cinfo.output_width * APP_FB_BPP;
     if (kStatus_Success != g_dc.ops->setLayerConfig(&g_dc, 0, &fbInfo))
     {
